+++ scaffolds/icestark-framework-app/README.md
@@ -1,11 +1,7 @@
 # Vite + icestark + vue3 框架应用
 
-Demo for building icestark micro applications created by Vite.
+Demo for building icestark framework application created by Vite.
 
 ## Online
 
-https://icestark-vite-vue.vercel.app/
-
-
-[Edit in stackBlitz](https://stackblitz.com/edit/github-kpsfcv)
-
+https://icestark-framework-vue3-x.vercel.app/
