+++ scaffolds/icestark-framework-app/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite App</title>
+    <title>icestark framework + vue3</title>
   </head>
   <body>
     <div id="app"></div>
